Separated measure report comparison utility from data
